# PARA PARCIAL/modelo-examen-api-elecciones-405708/docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'Parcial 2 - Elecciones 2023'","description":"'Parcial 2 - Elecciones 2023'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1'"},"servers":[{"url":"http://localhost:8081","description":"'Parcial 2 - Elecciones 2023'"}],"paths":{"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/elecciones/resultados":{"get":{"tags":["elecciones-controller"],"operationId":"getResultadosGenerales","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultadosDTO"}}}}}}},"/elecciones/distritos":{"get":{"tags":["elecciones-controller"],"operationId":"getDistritos","parameters":[{"name":"distrito_name","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DistritoDTO"}}}}}}}},"/elecciones/distritos/{id}":{"get":{"tags":["elecciones-controller"],"operationId":"getDistritoById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DistritoDTO"}}}}}}},"/elecciones/distritos/{id}/secciones":{"get":{"tags":["elecciones-controller"],"operationId":"getSeccionesByDistrito","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeccionDTO"}}}}}}}},"/elecciones/distritos/{id}/resultados":{"get":{"tags":["elecciones-controller"],"operationId":"getResultadosByDistrito","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultadoDistritoDTO"}}}}}}},"/elecciones/distritos/{id}/cargos":{"get":{"tags":["elecciones-controller"],"operationId":"getCargosDistrito","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CargoDTO"}}}}}}}},"/elecciones/distritos/{distritoId}/secciones/{seccionId}":{"get":{"tags":["elecciones-controller"],"operationId":"getSeccionByIdAndDistritoId","parameters":[{"name":"distritoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"seccionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SeccionDTO"}}}}}}},"/elecciones/distritos/{distritoId}/cargos/{cargoId}":{"get":{"tags":["elecciones-controller"],"operationId":"getCargoByIdAndDistritoId","parameters":[{"name":"distritoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"cargoId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CargoDTO"}}}}}}}},"components":{"schemas":{"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"ResultadoAgrupacionDTO":{"type":"object","properties":{"nombre":{"type":"string"},"posicion":{"type":"integer","format":"int32"},"votos":{"type":"integer","format":"int32"},"porcentaje":{"type":"string"}}},"ResultadoDistritoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"secciones":{"type":"array","items":{"type":"string"}},"votos_escrutados":{"type":"integer","format":"int32"},"porcentaje_padron_nacional":{"type":"number"},"agrupacion_ganadora":{"type":"string"},"resultados_agrupaciones":{"type":"array","items":{"$ref":"#/components/schemas/ResultadoAgrupacionDTO"}}}},"ResultadosDTO":{"type":"object","properties":{"distritos":{"type":"array","items":{"type":"string"}},"votos_escrutados":{"type":"integer","format":"int32"},"agrupacion_ganadora":{"type":"string"},"resultados_nacionales":{"type":"array","items":{"$ref":"#/components/schemas/ResultadoAgrupacionDTO"}},"resultados_distritos":{"type":"array","items":{"$ref":"#/components/schemas/ResultadoDistritoDTO"}}}},"DistritoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"}}},"SeccionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"}}},"CargoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"}}}}}}